requirements.txt
playwright==1.42.0
pytest==8.0.0
pytest-html==4.1.1
python-dotenv-==1.0.0